Create a new module for the ellipsis checker
